refactor

+++ fsoopify/serialize_ctx.py
@@ -24,7 +24,8 @@ class Context:
         self._atomic = atomic
         # states:
         self._lock_ctx = None
-        self._fp = None
+        self._openfd_ctx = None
+        self._fd = None
 
     def __enter__(self):
         def _read_data_from(fp):
@@ -36,11 +37,13 @@ class Context:
             is_exists = self._file_info.is_file()
 
             if self._atomic:
-                self._fp = self._file_info.open('wb', atomic=True)
+                self._openfd_ctx = self._file_info.open('wb', atomic=True)
             else:
-                self._fp = self._file_info.open_or_create('r+b')
+                self._openfd_ctx = self._file_info.open_or_create('r+b')
 
-            self._lock_ctx = self._lock(self._file_info, self._fp)
+            self._fd = self._openfd_ctx.__enter__()
+
+            self._lock_ctx = self._lock(self._file_info, self._fd)
             self._lock_ctx.__enter__()
 
             if is_exists:
@@ -48,8 +51,8 @@ class Context:
                     with self._file_info.open_for_read_bytes() as fp:
                         _read_data_from(fp)
                 else:
-                    self._fp.seek(0)
-                    _read_data_from(self._fp)
+                    self._fd.seek(0)
+                    _read_data_from(self._fd)
 
         except Exception:
             self._cleanup()
@@ -62,31 +65,33 @@ class Context:
             self._lock_ctx.__exit__(*sys.exc_info())
             self._lock_ctx = None
 
-        if self._fp is not None:
-            self._fp.__exit__(*sys.exc_info())
-            self._fp = None
+        if self._fd is not None:
+            assert self._openfd_ctx is not None
+            self._fd = None
+
+        if self._openfd_ctx is not None:
+            self._openfd_ctx.__exit__(*sys.exc_info())
+            self._openfd_ctx = None
 
     def __exit__(self, exc_type, exc_val, exc_tb):
+        remove_file = False
         try:
-            if exc_val is None and self.save_on_exit:
-                assert self._fp is not None
+            # only save if no exceptions:
+            if exc_val is None and self.save_on_exit and self._fd:
+                self._fd.seek(0)
                 if self.data is None:
-                    self._fp.__exit__(exc_type, exc_val, exc_tb)
-                    if self._file_info.is_file():
-                        self._file_info.delete()
-
+                    remove_file = True
                 else:
                     buf = self._serializer.dumpb(self.data, options={
                         'origin_kwargs': self._dump_kwargs
                     })
-                    self._fp.seek(0)
-                    self._fp.write(buf)
-                    self._fp.truncate()
-                    self._fp.__exit__(exc_type, exc_val, exc_tb)
-                self._fp = None
-
+                    self._fd.write(buf)
+                self._fd.truncate()
         finally:
             self._cleanup()
+
+        if remove_file and self._file_info.is_file():
+            self._file_info.delete()
 
 @contextmanager
 def lock_with_nop(fi, fp):
